test restructure and fixed typo
